requirements.txt
google-generativeai==0.8.1
fastapi==0.115.0
protobuf==5.28.2
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.30.6
